// api/types/paths/orders/v0/orders/{orderId}/orderItems/buyerInfo.types.ts
export type HTTP_GET = (
  $: OmitValueWhenNever<{
    query: { NextToken?: string };
    path: { orderId: string };
    header: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderItemsBuyerInfoResponse;
          };
        };
      };
      400: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderItemsBuyerInfoResponse;
          };
        };
      };
      403: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderItemsBuyerInfoResponse;
          };
        };
      };
      404: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderItemsBuyerInfoResponse;
          };
        };
      };
      429: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderItemsBuyerInfoResponse;
          };
        };
      };
      500: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderItemsBuyerInfoResponse;
          };
        };
      };
      503: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderItemsBuyerInfoResponse;
          };
        };
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
  }>
) =>
  | {
      status: 200;
      contentType?: 'application/json';
      body?: GetOrderItemsBuyerInfoResponse;
    }
  | {
      status: 400;
      contentType?: 'application/json';
      body?: GetOrderItemsBuyerInfoResponse;
    }
  | {
      status: 403;
      contentType?: 'application/json';
      body?: GetOrderItemsBuyerInfoResponse;
    }
  | {
      status: 404;
      contentType?: 'application/json';
      body?: GetOrderItemsBuyerInfoResponse;
    }
  | {
      status: 429;
      contentType?: 'application/json';
      body?: GetOrderItemsBuyerInfoResponse;
    }
  | {
      status: 500;
      contentType?: 'application/json';
      body?: GetOrderItemsBuyerInfoResponse;
    }
  | {
      status: 503;
      contentType?: 'application/json';
      body?: GetOrderItemsBuyerInfoResponse;
    }
  | { status: 415; contentType: 'text/plain'; body: string }
  | COUNTERFACT_RESPONSE
  | { ALL_REMAINING_HEADERS_ARE_OPTIONAL: COUNTERFACT_RESPONSE };